requirements.txt
anthropic>=0.40.0
twikit @ git+https://github.com/pothitos/twikit.git@patch-1
gspread>=6.0.0
google-auth>=2.30.0
python-dotenv>=1.0.0
